Ensure the segments array is updated
